Fix: "엔티티 필드값 수정"

+++ src/main/java/com/app/domain/categorizedProblem/entity/CategorizedProblem.java
@@ -15,7 +15,7 @@ import javax.persistence.*;
 public class CategorizedProblem extends BaseEntity {
     @Id
     @Column(name = "CATEGORIZED_PROBLEM_ID")
-    private Integer categorizedProblemId;
+    private Long categorizedProblemId;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "CATEGORY_ID", nullable = false)
